backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
python-decouple==3.8
requests==2.31.0
Pillow==10.4.0
tensorflow==2.19.0
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.5.1
openai==1.3.7
python-dotenv==1.0.0
google-generativeai==0.3.2
gunicorn==21.2.0
whitenoise[brotli]==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1